Add tests and refactor lib
