+++ package.json
@@ -5,7 +5,7 @@
   "main": "index.js",
   "scripts": {
     "start": "node index.js",
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "tap test/**/test-*.js"
   },
   "repository": {
     "type": "git",
@@ -27,5 +27,8 @@
   "dependencies": {
     "fluent-ffmpeg": "^2.1.0",
     "simple-git": "^1.61.0"
+  },
+  "devDependencies": {
+    "tap": "^8.0.0"
   }
 }
